requirements.txt
numpy
psycopg[binary]
psycopg2-binary
